requirements.txt
python-telegram-bot[webhooks]==21.6
aiohttp==3.9.5